+++ .gitignore
@@ -4,6 +4,7 @@ target/
 !.mvn/wrapper/maven-wrapper.jar
 !**/src/main/**/target/
 !**/src/test/**/target/
+/log/
 
 ### STS ###
 .apt_generated
