+++ src/main/java/com/example/application/model/Material.java
@@ -9,6 +9,7 @@ import javax.persistence.*;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
+import java.util.Date;
 import java.util.Objects;
 
 @Entity
@@ -28,6 +29,25 @@ public class Material extends AbstractNamedEntity implements HasId {
   @Size(min = 2, max = 120)
   private String characteristics;
 
+  @Column(name = "hasFactoryMarriage")
+  private boolean hasFactoryMarriage;
+
+  @Column(name = "marriageDetectionDate")
+  @Size(min = 2, max = 20)
+  private Date marriageDetectionDate;
+
+  @Column(name = "marriageDetectedBy")
+  @Size(min = 2, max = 128)
+  private String marriageDetectedBy;
+
+  @Column(name = "marriageDescription")
+  @Size(min = 2, max = 120)
+  private String marriageDescription;
+
+  @Column(name = "marriagePhotoUrl")
+  @Size(min = 2, max = 150)
+  private String marriagePhotoUrl;
+
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "ticket_id")
   @OnDelete(action = OnDeleteAction.CASCADE)
@@ -37,6 +57,12 @@ public class Material extends AbstractNamedEntity implements HasId {
     this.name = name;
     this.quantity = quantity;
     this.characteristics = characteristics;
+
+    boolean hasFactoryMarriage = false;
+    Date marriageDetectionDate;
+    String marriageDetectedBy;
+    String marriageDescription;
+    String marriagePhotoUrl;
   }
 
   @Override
@@ -68,6 +94,46 @@ public class Material extends AbstractNamedEntity implements HasId {
     this.characteristics = characteristics;
   }
 
+  public boolean isHasFactoryMarriage() {
+    return hasFactoryMarriage;
+  }
+
+  public void setHasFactoryMarriage(boolean hasFactoryMarriage) {
+    this.hasFactoryMarriage = hasFactoryMarriage;
+  }
+
+  public Date getMarriageDetectionDate() {
+    return marriageDetectionDate;
+  }
+
+  public void setMarriageDetectionDate(Date marriageDetectionDate) {
+    this.marriageDetectionDate = marriageDetectionDate;
+  }
+
+  public String getMarriageDetectedBy() {
+    return marriageDetectedBy;
+  }
+
+  public void setMarriageDetectedBy(String marriageDetectedBy) {
+    this.marriageDetectedBy = marriageDetectedBy;
+  }
+
+  public String getMarriageDescription() {
+    return marriageDescription;
+  }
+
+  public void setMarriageDescription(String marriageDescription) {
+    this.marriageDescription = marriageDescription;
+  }
+
+  public String getMarriagePhotoUrl() {
+    return marriagePhotoUrl;
+  }
+
+  public void setMarriagePhotoUrl(String marriagePhotoUrl) {
+    this.marriagePhotoUrl = marriagePhotoUrl;
+  }
+
   public Ticket getTicket() {
     return ticket;
   }
@@ -76,28 +142,28 @@ public class Material extends AbstractNamedEntity implements HasId {
     this.ticket = ticket;
   }
 
+  public void detectFactoryMarriage(Date marriageDetectionDate, String marriageDetectedBy,
+                                    String marriageDescription, String marriagePhotoUrl) {
+    setHasFactoryMarriage(true);
+    setMarriageDetectionDate(marriageDetectionDate);
+    setMarriageDetectedBy(marriageDetectedBy);
+    setMarriageDescription(marriageDescription);
+    setMarriagePhotoUrl(marriagePhotoUrl);
+  }
+
   @Override
   public String toString() {
     return "Material{" +
             "id=" + id +
-            "quantity=" + quantity +
+            ", quantity=" + quantity +
             ", characteristics='" + characteristics + '\'' +
+            ", hasFactoryMarriage=" + hasFactoryMarriage +
+            ", marriageDetectionDate=" + marriageDetectionDate +
+            ", marriageDetectedBy='" + marriageDetectedBy + '\'' +
+            ", marriageDescription='" + marriageDescription + '\'' +
+            ", marriagePhotoUrl='" + marriagePhotoUrl + '\'' +
             ", ticket=" + ticket +
-            ", id=" + id +
             ", name='" + name + '\'' +
             '}';
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) return true;
-    if (o == null || getClass() != o.getClass()) return false;
-    Material material = (Material) o;
-    return id.equals(material.id);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(id);
   }
 }
