+++ src/main/java/com/example/application/model/Role.java
@@ -4,7 +4,11 @@ import org.springframework.security.core.GrantedAuthority;
 
 public enum Role implements GrantedAuthority {
   ADMINISTRATOR,
-  USER;
+  USER,
+
+  SUPPLIER,
+
+  FOREMAN;
 
   //    https://stackoverflow.com/a/19542316/548473
   @Override
