+++ src/main/java/com/example/application/repository/datajpa/CrudMaterialRepository.java
@@ -9,14 +9,14 @@ import org.springframework.data.repository.query.Param;
 public interface CrudMaterialRepository extends JpaRepository<Material, Integer> {
 
     @Modifying
-    @Query("INSERT INTO Material (name, quantity, characteristics) SELECT mt.name, mt.quantity, mt.characteristics FROM MaterialTo mt")
+//    @Query("INSERT INTO Material (name, quantity, characteristics) SELECT mt.name, mt.quantity, mt.characteristics FROM MaterialTo mt")
     Material save(Material entity);
 
     @Modifying
     @Query("DELETE FROM Material m WHERE m.id=:id AND m.ticket.id=:ticketId")
     int delete(@Param("id") int id, @Param("ticketId") int ticketId);
 
-    @Query("SELECT t.name as 'Название заявки', t.creationDate as 'Дата создания', t.status as 'Статус', m.name as 'Название материала', m.quantity as 'Количество материала'\n" +
-            "FROM materials m RIGHT JOIN tickets t ON m.ticket_id = t.id WHERE m.id=:id AND m.name=:name")
-    Material findByName(@Param("id") int id, @Param("name") String name);
+//    @Query("SELECT t.name as 'Название заявки', t.creationDate as 'Дата создания', t.status as 'Статус', m.name as 'Название материала', m.quantity as 'Количество материала'\n" +
+//            "FROM materials m RIGHT JOIN tickets t ON m.ticket_id = t.id WHERE m.id=:id AND m.name=:name")
+//    Material findByName(@Param("id") int id, @Param("name") String name);
 }
