+++ src/main/java/com/example/application/repository/datajpa/CrudTicketRepository.java
@@ -13,17 +13,17 @@ import java.util.List;
 public interface CrudTicketRepository extends JpaRepository<Ticket, Integer> {
 
     @Modifying
-    @Query("INSERT INTO Ticket (user_id, name, status) SELECT tt.user_id, tt.name, tt.status FROM TicketTo tt")
+//    @Query("INSERT INTO Ticket (user_id, name, status) SELECT tt.user_id, tt.name, tt.status FROM TicketTo tt")
     Ticket save(Ticket entity);
 
     @Modifying
-    @Query("DELETE FROM Ticket t WHERE t.id=:id AND t.user.id=:id")
+    @Query("DELETE FROM Ticket t WHERE t.id=:id AND t.user.id=:user_id")
     int delete(@Param("id") int id, @Param("user_id") int user_id);
 
     @Query("SELECT t FROM Ticket t WHERE t.id=:id")
     Ticket findById(@Param("id") int id);
 
-    @EntityGraph(attributePaths = {"tickets_materials"}, type = EntityGraph.EntityGraphType.LOAD)
-    @Query("SELECT t FROM Ticket t INNER JOIN t.materials m WHERE t.id=:id AND t.user.id=:userId AND t.id=m.ticket.id")
-    Ticket getWithMaterial(@Param("id") int id, @Param("userId") int userId, List<Material> materialList);
+//    @EntityGraph(attributePaths = {"tickets_materials"}, type = EntityGraph.EntityGraphType.LOAD)
+//    @Query("SELECT t FROM Ticket t INNER JOIN t.materials m WHERE t.id=:id AND t.user.id=:userId AND t.id=m.ticket.id")
+//    Ticket getWithMaterial(@Param("id") int id, @Param("userId") int userId, List<Material> materialList);
 }
