+++ src/main/java/com/example/application/util/UserUtil.java
@@ -1,32 +1,32 @@
-package com.example.application.util;
-
-import org.springframework.security.crypto.password.PasswordEncoder;
-import com.example.application.model.User;
-import com.example.application.to.UserTo;
-
-import java.util.Date;
-
-public class UserUtil {
-
-
-    public static User createNewFromTo(UserTo userTo) {
-        return new User(null, userTo.getName(), userTo.getEmail().toLowerCase(), userTo.getPassword(), userTo.getRole(), new Date());
-    }
-
-    public static UserTo asTo(User user) {
-        return new UserTo(user.getId(), user.getName(), user.getEmail(), user.getPassword(), user.getRoles());
-    }
-
-    public static User updateFromTo(User user, UserTo userTo) {
-        user.setName(userTo.getName());
-        user.setEmail(userTo.getEmail().toLowerCase());
-        user.setPassword(userTo.getPassword());
-        return user;
-    }
-
-    public static User prepareToSave(User user, PasswordEncoder passwordEncoder) {
-        user.setPassword(passwordEncoder.encode(user.getPassword()));
-        user.setEmail(user.getEmail().toLowerCase());
-        return user;
-    }
-}
+//package com.example.application.util;
+//
+//import org.springframework.security.crypto.password.PasswordEncoder;
+//import com.example.application.model.User;
+//import com.example.application.to.UserTo;
+//
+//import java.util.Date;
+//
+//public class UserUtil {
+//
+//
+//    public static User createNewFromTo(UserTo userTo) {
+//        return new User(null, userTo.getName(), userTo.getEmail().toLowerCase(), userTo.getPassword(), userTo.getRole(), new Date());
+//    }
+//
+//    public static UserTo asTo(User user) {
+//        return new UserTo(user.getId(), user.getName(), user.getEmail(), user.getPassword(), user.getRoles());
+//    }
+//
+//    public static User updateFromTo(User user, UserTo userTo) {
+//        user.setName(userTo.getName());
+//        user.setEmail(userTo.getEmail().toLowerCase());
+//        user.setPassword(userTo.getPassword());
+//        return user;
+//    }
+//
+//    public static User prepareToSave(User user, PasswordEncoder passwordEncoder) {
+//        user.setPassword(passwordEncoder.encode(user.getPassword()));
+//        user.setEmail(user.getEmail().toLowerCase());
+//        return user;
+//    }
+//}
