+++ src/main/java/com/example/application/web/RootController.java
@@ -1,0 +1,53 @@
+package com.example.application.web;
+
+import com.example.application.service.TicketService;
+import com.example.application.service.UserService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+
+import javax.servlet.http.HttpServletRequest;
+
+@Controller
+public class RootController {
+    private static final Logger log = LoggerFactory.getLogger(RootController.class);
+
+    @Autowired
+    private UserService userService;
+
+    @Autowired
+    private TicketService ticketService;
+
+    @GetMapping("/")
+    public String root() {
+        log.info("root");
+        return "index";
+    }
+
+    @GetMapping("/users")
+    public String getUsers(Model model) {
+        log.info("users");
+        model.addAttribute("users", userService.getAll());
+        return "users";
+    }
+
+    @PostMapping("/users")
+    public String setUser(HttpServletRequest request) {
+        int userId = Integer.parseInt(request.getParameter("userId"));
+        log.info("setUser {}", userId);
+        SecurityUtil.setAuthUserId(userId);
+        return "redirect:tickets";
+    }
+
+    @GetMapping("/tickets")
+    public String getMeals(Model model) {
+        log.info("tickets");
+        model.addAttribute("tickets",
+                ticketService.getAll(SecurityUtil.authUserId()));
+        return "tickets";
+    }
+}
