+++ src/main/java/com/example/application/web/SecurityUtil.java
@@ -1,0 +1,18 @@
+package com.example.application.web;
+
+public class SecurityUtil {
+
+    private static int id = 1;
+
+    private SecurityUtil() {
+    }
+
+    public static int authUserId() {
+        return id;
+    }
+
+    public static void setAuthUserId(int id) {
+        SecurityUtil.id = id;
+    }
+
+}
