--- src/main/resources/application.properties
@@ -1,6 +1,0 @@
-spring.jpa.hibernate.ddl-auto=update
-spring.datasource.url=jdbc:mysql://gadgi95.beget.tech:3306/gadgi95_applicat?autoReconnect=true&serverTimezone=UTC&useSSL=false
-spring.datasource.username=gadgi95_applicat
-spring.datasource.password=Qwerty!23
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-#spring.jpa.show-sql: false
