+++ src/main/resources/db/application.properties
@@ -1,0 +1,12 @@
+spring.jpa.hibernate.ddl-auto=update
+database.url=jdbc:mysql://gadgi95.beget.tech:3306/gadgi95_applicat?autoReconnect=true&serverTimezone=UTC&useSSL=false
+database.username=gadgi95_applicat
+database.password=Qwerty!23
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+#spring.jpa.show-sql: false
+
+database.init=true
+jpa.showSql=true
+jdbc.initLocation=classpath:db/initDB.sql
+hibernate.format_sql=true
+hibernate.use_sql_comments=true
