+++ src/main/resources/spring/spring-security.xml
@@ -1,43 +1,43 @@
-<beans:beans xmlns="http://www.springframework.org/schema/security"
-             xmlns:beans="http://www.springframework.org/schema/beans"
-             xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-             xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
-	http://www.springframework.org/schema/security http://www.springframework.org/schema/security/spring-security.xsd">
+<!--<beans:beans xmlns="http://www.springframework.org/schema/security"-->
+<!--             xmlns:beans="http://www.springframework.org/schema/beans"-->
+<!--             xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"-->
+<!--             xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd-->
+<!--	http://www.springframework.org/schema/security http://www.springframework.org/schema/security/spring-security.xsd">-->
 
-    <http pattern="/resources/**" security="none"/>
-    <http pattern="/webjars/**" security="none"/>
+<!--    <http pattern="/resources/**" security="none"/>-->
+<!--    <http pattern="/webjars/**" security="none"/>-->
 
-    <http pattern="/rest/**" use-expressions="true" name="restSecurityFilterChain" create-session="stateless">
-        <http-basic/>
-        <intercept-url pattern="/rest/admin/**" access="hasRole('ADMIN')"/>
-        <intercept-url pattern="/rest/profile" method="POST" access="isAnonymous()"/>
-        <intercept-url pattern="/**" access="isAuthenticated()"/>
-        <csrf disabled="true"/>
-    </http>
+<!--    <http pattern="/rest/**" use-expressions="true" name="restSecurityFilterChain" create-session="stateless">-->
+<!--        <http-basic/>-->
+<!--        <intercept-url pattern="/rest/admin/**" access="hasRole('ADMIN')"/>-->
+<!--        <intercept-url pattern="/rest/profile" method="POST" access="isAnonymous()"/>-->
+<!--        <intercept-url pattern="/**" access="isAuthenticated()"/>-->
+<!--        <csrf disabled="true"/>-->
+<!--    </http>-->
 
-    <beans:bean class="org.springframework.security.crypto.factory.PasswordEncoderFactories" id="passwordEncoder" factory-method="createDelegatingPasswordEncoder"/>
+<!--    <beans:bean class="org.springframework.security.crypto.factory.PasswordEncoderFactories" id="passwordEncoder" factory-method="createDelegatingPasswordEncoder"/>-->
 
-    <http>
-        <intercept-url pattern="/login" access="permitAll()"/>
+<!--    <http>-->
+<!--        <intercept-url pattern="/login" access="permitAll()"/>-->
 
-        <intercept-url pattern="/swagger-ui.html" access="permitAll()"/>
-        <intercept-url pattern="/swagger-resources/**" access="permitAll()"/>
-        <intercept-url pattern="/v2/api-docs/**" access="permitAll()"/>
+<!--        <intercept-url pattern="/swagger-ui.html" access="permitAll()"/>-->
+<!--        <intercept-url pattern="/swagger-resources/**" access="permitAll()"/>-->
+<!--        <intercept-url pattern="/v2/api-docs/**" access="permitAll()"/>-->
 
-        <intercept-url pattern="/profile/register" access="isAnonymous()"/>
+<!--        <intercept-url pattern="/profile/register" access="isAnonymous()"/>-->
 
-        <intercept-url pattern="/admin/**" access="hasRole('ADMIN')"/>
-        <intercept-url pattern="/**" access="isAuthenticated()"/>
+<!--        <intercept-url pattern="/admin/**" access="hasRole('ADMIN')"/>-->
+<!--        <intercept-url pattern="/**" access="isAuthenticated()"/>-->
 
-        <form-login login-page="/login" default-target-url="/meals"
-                    authentication-failure-url="/login?error=true"
-                    login-processing-url="/spring_security_check"/>
-        <logout logout-success-url="/login"/>
-    </http>
+<!--        <form-login login-page="/login" default-target-url="/meals"-->
+<!--                    authentication-failure-url="/login?error=true"-->
+<!--                    login-processing-url="/spring_security_check"/>-->
+<!--        <logout logout-success-url="/login"/>-->
+<!--    </http>-->
 
-    <authentication-manager>
-        <authentication-provider user-service-ref="userService">
-            <password-encoder ref="passwordEncoder"/>
-        </authentication-provider>
-    </authentication-manager>
-</beans:beans>
+<!--    <authentication-manager>-->
+<!--        <authentication-provider user-service-ref="userService">-->
+<!--            <password-encoder ref="passwordEncoder"/>-->
+<!--        </authentication-provider>-->
+<!--    </authentication-manager>-->
+<!--</beans:beans>-->
