+++ src/main/resources/tomcat/context.xml
@@ -1,0 +1,57 @@
+<?xml version='1.0' encoding='utf-8'?>
+<!--
+  Licensed to the Apache Software Foundation (ASF) under one or more
+  contributor license agreements.  See the NOTICE file distributed with
+  this work for additional information regarding copyright ownership.
+  The ASF licenses this file to You under the Apache License, Version 2.0
+  (the "License"); you may not use this file except in compliance with
+  the License.  You may obtain a copy of the License at
+
+      http://www.apache.org/licenses/LICENSE-2.0
+
+  Unless required by applicable law or agreed to in writing, software
+  distributed under the License is distributed on an "AS IS" BASIS,
+  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  See the License for the specific language governing permissions and
+  limitations under the License.
+-->
+<!-- The contents of this file will be loaded for each web application -->
+<Context>
+
+    <!-- Default set of monitored resources. If one of these changes, the    -->
+    <!-- web application will be reloaded.                                   -->
+    <WatchedResource>WEB-INF/web.xml</WatchedResource>
+    <WatchedResource>${catalina.base}/conf/web.xml</WatchedResource>
+
+    <!-- Uncomment this to disable session persistence across Tomcat restarts -->
+    <!--
+    <Manager pathname="" />
+    -->
+
+    <!-- Uncomment this to enable Comet connection tacking (provides events
+         on session expiration as well as webapp lifecycle) -->
+    <!--
+    <Valve className="org.apache.catalina.valves.CometConnectionManagerValve" />
+    -->
+    <Resource name="jdbc/application"
+              auth="Container"
+              type="javax.sql.DataSource"
+              url="jdbc:mysql://gadgi95.beget.tech:3306/gadgi95_applicat?autoReconnect=true"
+              username="gadgi95_applicat"
+              password="Qwerty!23"
+              driverClassName="com.mysql.cj.jdbc.Driver"
+              validationQuery="SELECT 1"
+              maxTotal="10"
+              minIdle="2"
+              maxWaitMillis="20000"
+              initialSize="2"
+              maxIdle="5"
+              testOnBorrow="true"
+              removeAbandonedOnBorrow="true"
+              testWhileIdle="true"/>
+    <!--
+        username="wegxlfzjjgxaxy"
+        password="SSQyKKE_e93kiUCR-ehzMcKCxZ"
+        url="jdbc:postgresql://ec2-54-217-202-110.eu-west-1.compute.amazonaws.com:5432/dehm6lvm8bink0?ssl=true&amp;sslfactory=org.postgresql.ssl.NonValidatingFactory"/>
+    -->
+</Context>
