+++ src/main/webapp/WEB-INF/jsp/fragments/headTag.jsp
@@ -6,9 +6,5 @@
     <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
     <title><spring:message code="app.title"/></title>
     <base href="${pageContext.request.contextPath}/"/>
-
-    <link rel="stylesheet" href="resources/css/style.css?v=2">
-    <link rel="stylesheet" href="webjars/bootstrap/4.6.1/css/bootstrap.min.css">
-    <link rel="stylesheet" href="webjars/noty/3.1.4/demo/font-awesome/css/font-awesome.min.css">
-    <link rel="shortcut icon" href="resources/images/home.png">
+    <link rel="stylesheet" href="resources/css/style.css">
 </head>
