+++ src/test/java/com/example/application/core/MaterialTest.java
@@ -6,7 +6,7 @@ import com.example.application.model.Material;
 import org.junit.jupiter.api.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
+//import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.stereotype.Component;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
@@ -16,7 +16,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 @Component
 @TestPropertySource(locations = "classpath:application-context.xml")
 @RunWith(SpringRunner.class)
-@SpringBootTest(classes = { Material.class })
+//@SpringBootTest(classes = { Material.class })
 public class MaterialTest {
 
     @Autowired
