+++ src/test/java/com/example/application/repository/AbstractServiceTest.java
@@ -5,7 +5,7 @@ import org.springframework.test.context.jdbc.SqlConfig;
 import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
 
 @SpringJUnitConfig(locations = {
-        "classpath:application.properties"
+        "classpath:db/application.properties"
 })
 @Sql(scripts = "classpath:db/populateDB.sql", config = @SqlConfig(encoding = "UTF-8"))
 public abstract class AbstractServiceTest {
